更新 User Group Controller

+++ examsystem/src/main/java/com/monktiger/examsystem/controller/home/GroupController.java
@@ -2,45 +2,86 @@ package com.monktiger.examsystem.controller.home;
 
 import com.monktiger.examsystem.cache.JedisUtil;
 import com.monktiger.examsystem.entity.Group;
+import com.monktiger.examsystem.mapper.GroupMapper;
 import com.monktiger.examsystem.service.impl.GroupServiceImpl;
+import com.monktiger.examsystem.util.MD5Util;
 import com.monktiger.examsystem.utils.resultHander.CommonErrorEnum;
 import com.monktiger.examsystem.utils.resultHander.ResponseBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 @RequestMapping("/group")
 @RestController
+@CrossOrigin
 public class GroupController {
 
     @Autowired
     private GroupServiceImpl groupService;
+
     @Autowired
+    private GroupMapper groupMapper;
+
     private JedisUtil jedisUtil;
+
+    @Autowired
+    private JedisUtil.Keys jedisUtilKeys;
+
+    @Autowired
+    private JedisUtil.Strings jedistUtilStrings;
+
+    private Map<String,Object> modelMap = new HashMap<>();
+
 
 
     /**
      * 创建组
-     * @param groupName
+     * @param name
      * @return
      * @throws Exception
      */
     @RequestMapping("/create")
-    public Object create(String groupName) throws Exception {
-        String openId = jedisUtil.STRINGS.get("openId");
-        Group group = new Group(groupName);
-        group.setOpenId(openId);
-        //groupId为6位字符组合所以我通过md5+时间戳生成6位随机值作为groupId
+    public Map<String,Object> create(String name, HttpServletRequest request) throws Exception {
+        String token = request.getHeader("token");
+        if(token!=null&&jedisUtilKeys.exists("token")){
 
-        if(groupService.insertGroup(group)==1){
-            return new ResponseBean<>(true,"创建 group 成功", CommonErrorEnum.SUCCESS_OPTION);
+            String openId = jedistUtilStrings.get("token");
+//            String openId = "openId";
+            Group group = new Group(name);
+            group.setOpenId(openId);
+            //groupId为6位字符组合所以我通过md5+时间戳生成6位随机值作为groupId
+
+            /**
+             * 创建唯一 字符串作为 groupId
+             */
+            String groupId = MD5Util.getMd5(String.valueOf(System.currentTimeMillis()),6);;
+            group.setGroupId(groupId);
+            if(groupService.insertGroup(group)==1){
+                modelMap.put("groupId",groupId);
+                modelMap.put("status",1);
+                modelMap.put("msg","创建成功");
+            }else{
+                modelMap.put("groupId",groupId);
+                modelMap.put("status",0);
+                modelMap.put("msg","创建失败");
+            }
+
         }else{
-            return new ResponseBean<>(false,"创建 group 失败",CommonErrorEnum.FAILED_QUESTION);
+            modelMap.put("state",0);
+            modelMap.put("msg","请登录后操作");
         }
+        return modelMap;
     }
+
 
     /**
      * 加入组
@@ -48,19 +89,32 @@ public class GroupController {
      * @return
      * @throws Exception
      */
-    @RequestMapping("/join")
-    public Object join(String groupId)throws Exception{
-        String openId = jedisUtil.STRINGS.get("openId");
-        Group group = new Group(groupId);
-        group.setOpenId(openId);
-        int state = groupService.joinGroup(group);
-        if(state==1){
-            return new ResponseBean<>(true,"加入成功",CommonErrorEnum.SUCCESS_OPTION);
-        }else if(state==0){
-            return new ResponseBean<>(false,"",CommonErrorEnum.FAILED_QUESTION);
+    @RequestMapping(value="/join",method = RequestMethod.GET)
+    public Map<String,Object> join(String groupId, HttpServletRequest request)throws Exception{
+        String token = request.getHeader("token");
+
+        if(token!=null&&jedisUtilKeys.exists("token"))
+        {
+            String openId = token;
+            Group group = new Group(groupId);
+            group.setGroupId(groupId);
+            group.setOpenId(openId);
+            int state = groupService.joinGroup(group);
+            if(state==1){
+                modelMap.put("status",1);
+                modelMap.put("msg","成功加入");
+//            return new ResponseBean<>(true,"加入成功",CommonErrorEnum.SUCCESS_OPTION);
+            }else {
+                modelMap.put("status", 0);
+                modelMap.put("msg", "失败加入");
+//            return new ResponseBean<>(false,"",CommonErrorEnum.FAILED_QUESTION);
+            }
         }else{
-            return new ResponseBean<>(false,"", CommonErrorEnum.FAILED_QUESTION);
+            modelMap.put("state",0);
+            modelMap.put("msg","请登录后操作");
         }
+
+        return modelMap;
     }
 
     /**
@@ -70,31 +124,50 @@ public class GroupController {
      * @throws Exception
      */
     @RequestMapping("/leave")
-    public Object leave(String groupId)throws Exception{
-        String openId = jedisUtil.STRINGS.get("openId");
-        Group group = new Group(groupId,openId);
-        int state = groupService.deleteGroup(group);
-        if(state == 1){
-            return new ResponseBean<>(true,"退组成功",CommonErrorEnum.SUCCESS_OPTION);
+    public Map<String,Object> leave(String groupId,HttpServletRequest request)throws Exception{
+        String token = request.getHeader("token");
+        if(token!=null&&jedisUtilKeys.exists("token"))
+        {
+            String openId = token;
+            Group group = new Group(groupId,"openId");
+            int state = groupService.deleteGroup(group);
+            if(state == 1){
+                modelMap.put("status",1);
+                modelMap.put("msg","退出成功");
+
+            }else{
+//            return new ResponseBean<>(false,"退组失败", CommonErrorEnum.FAILED_QUESTION);
+                modelMap.put("status",0);
+                modelMap.put("msg","退出失败");
+            }
         }else{
-            return new ResponseBean<>(false,"退组失败", CommonErrorEnum.FAILED_QUESTION);
+            modelMap.put("state",0);
+            modelMap.put("msg","请登录后操作");
         }
+        return modelMap;
     }
 
     /**
-     * 列出 自己相关组
+     * 列出 自己相关组 &
      * @return
      * @throws Exception
      */
-    @RequestMapping("/getList")
-    public Object getList()throws Exception{
-        String openId = jedisUtil.STRINGS.get("openId");
-        List<Group> groups = groupService.selectSelfGroup(openId);
-        if(groups!=null){
-            return new ResponseBean<>(true,groups, CommonErrorEnum.SUCCESS_OPTION);
+    @RequestMapping(value="/getList",method= RequestMethod.GET)
+    public Object getList(HttpServletRequest request,Integer type)throws Exception{
+        String token = request.getHeader("token");
+        if(token!=null&&jedisUtilKeys.exists("token"))
+        {
+            String openId = token;
+            List<Group> groups = groupMapper.selectSelfGroup(openId,type);
+            modelMap.put("groups",groups);
+            modelMap.put("state",1);
+            modelMap.put("msg","加入成功");
         }else{
-            return new ResponseBean<>(false,"空的",CommonErrorEnum.FAILED_QUESTION);
+            modelMap.put("state",0);
+            modelMap.put("msg","请登录后操作");
         }
+//        String openId = jedisUtil.STRINGS.get("openId");
+        return modelMap;
     }
 
 
@@ -109,7 +182,8 @@ public class GroupController {
      */
     @RequestMapping("/deleteMember")
     public Object deleteMember(String groupId,List<String> openIds)throws Exception{
-        String openId = jedisUtil.STRINGS.get("openId");
+//        String openId = jedisUtil.STRINGS.get("openId");
+        String  openId = "openId";
         Group aim_group = groupService.selectGroupById(groupId,openId);
         if(aim_group==null){
             return new ResponseBean<>(false,"查无此人",CommonErrorEnum.FAILED_QUESTION);
@@ -128,13 +202,25 @@ public class GroupController {
      * @return
      * @throws Exception
      */
-    @RequestMapping("/showMember")
-    public Object showMember(String groupId)throws Exception{
+    @RequestMapping(value="/showMember",method = RequestMethod.GET)
+    public Map<String,Object> showMember(String groupId,HttpServletRequest request)throws Exception{
+        String token = request.getHeader("token");
+        if(token!=null&&jedisUtilKeys.exists("token"))
+        {
+
+            modelMap.put("state",1);
+            modelMap.put("msg","加入成功");
+        }else{
+            modelMap.put("state",0);
+            modelMap.put("msg","请登录后操作");
+        }
         Group group = new Group();
         group.setGroupId(groupId);
         group.setStatus(1);
-        List<Group> groups = groupService.selectSelfGroup(groupId);
-        return new ResponseBean<>(true,groups,CommonErrorEnum.SUCCESS_OPTION);
+        List<Group> groups = groupMapper.selectByKeyState(groupId,0);
+        modelMap.put("groups",groups);
+        modelMap.put("state",1);
+        return modelMap;
     }
 
     /**
@@ -143,21 +229,33 @@ public class GroupController {
      * @return
      * @throws Exception
      */
-    @RequestMapping("/delete")
-    public Object delete(String groupId)throws Exception{
-        String openId = jedisUtil.STRINGS.get("openId");
-        Group group = groupService.selectGroupById(openId,groupId);
+    @RequestMapping("/quit")
+    public Map<String,Object> delete(String groupId,HttpServletRequest request)throws Exception{
+        String token = request.getHeader("token");
+        if(token!=null&&jedisUtilKeys.exists("token"))
+        {
+            //String openId = jedisUtil.STRINGS.get("openId");
+            String openId = token;
+            Group group = groupService.selectGroupById(openId,groupId);
 
-        /**
-         * 判断 执行操作的是组长 还是 组员
-         */
-        if(group.getStatus() == 0){
-            groupService.deleteGroup(new Group(groupId,openId));
-            return new ResponseBean(true,"推出小组成功");
+            /**
+             * 判断 执行操作的是组长 还是 组员
+             */
+            if(group.getStatus() == 1){
+                groupService.deleteGroup(new Group(groupId,openId));
+                modelMap.put("state",1);
+                modelMap.put("msg","退出小组成功");
+            }else{
+                groupService.dissolve(groupId);
+                modelMap.put("state",2);
+                modelMap.put("msg","解散小组成功");
+            }
         }else{
-            groupService.dissolve(groupId);
-            return new ResponseBean(true,"解散小组成功");
+            modelMap.put("state",0);
+            modelMap.put("msg","请登录后操作");
         }
+
+        return modelMap;
     }
